+++ webpack-packages/templating-plugin/prerender/index.js
@@ -1,7 +1,7 @@
 const puppeteer = require('puppeteer');
 
 module.exports = async (url, config) => {
-  const browser = await puppeteer.launch({headless: true, args: ['--no-sandbox']});
+  const browser = await puppeteer.launch({headless: true, args: ['--no-sandbox', '--disable-setuid-sandbox']});
   const page = await browser.newPage();
 
   page.evaluateOnNewDocument(`