+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0.4] - 2019-06-18
 ### [Added]
+- @aofl/cli - reporter option for build and test
+
 ### [Changed]
 - @aofl/templating-plugin - updated puppeteer and added --no-sandbox